backend/analytics/requirements.txt
Flask==2.0.3
google-cloud-datastore
